+++ code/OnionServer/src/main/java/com/onion/onionserver/repo/UserAuthRepo.java
@@ -3,5 +3,5 @@ package com.onion.onionserver.repo;
 import com.onion.onionserver.model.dao.UserAuth;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserAuthRepo extends CrudRepository<UserAuth, Long> {
+public interface UserAuthRepo extends CrudRepository<UserAuth, Integer> {
 }
